Refactor AudiocodesRealtimeBridge to TelephonyRealtimeBridge and introduce new module
This commit renames the AudiocodesRealtimeBridge to TelephonyRealtimeBridge across the codebase, updating all relevant import statements. Additionally, a new TelephonyRealtimeBridge module is introduced, providing a robust integration between AudioCodes WebSocket protocol and OpenAI's Realtime API. The new module includes comprehensive documentation, performance optimizations, and improved error handling for real-time audio streaming. Unit tests are also added to ensure functionality and reliability.
