+++ app/main.py
@@ -19,7 +19,7 @@ from fastapi import FastAPI, WebSocket
 
 from app.config.logging_config import configure_logging
 from app.websocket_manager import WebSocketManager
-from app.bot.audiocodes_realtime_bridge import bridge
+from app.bot.telephony_realtime_bridge import bridge
 
 # Load environment variables from .env file if it exists
 env_path = Path(".") / ".env"
