+++ tests/test_stream_handlers.py
@@ -21,7 +21,7 @@ from app.models.audiocodes_api import (
     UserStreamStopMessage,
     UserStreamStoppedResponse,
 )
-from app.bot.audiocodes_realtime_bridge import bridge
+from app.bot.telephony_realtime_bridge import bridge
 
 
 @pytest.mark.asyncio
